backend/requirements.txt
requests==2.32.4
beautifulsoup4==4.13.4
lxml==5.4.0
openai==1.91.0
tqdm==4.67.1
python-dotenv==1.0.1
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pydantic==2.11.7
asyncpg==0.30.0
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10 